+++ apps/miranum-forms/CHANGELOG.md
@@ -7,10 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.3] - 2023-07-03
+
 ### Changed
 
 * Update dependencies and migrate to Nx Version 16 (#246)
 
 
 
-[unreleased]: https://github.com/Miragon/miranum-ide/compare/release/v0.4.2...HEAD
+[unreleased]: https://github.com/Miragon/miranum-ide/compare/release/v0.4.3...HEAD
+[0.4.3]: https://github.com/Miragon/miranum-ide/compare/release/v0.4.2...0.4.3