+++ apps/miranum-modeler/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.3] - 2023-07-03
+
 ### Added
 
 * The modeler now supports DMN 1.3 (#219)
@@ -21,4 +23,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-[unreleased]: https://github.com/Miragon/miranum-ide/compare/release/v0.4.2...HEAD
+[unreleased]: https://github.com/Miragon/miranum-ide/compare/release/v0.4.3...HEAD
+[0.4.3]: https://github.com/Miragon/miranum-ide/compare/release/v0.4.2...0.4.3